Update PowerPoint.ExtraColors.md
Changes:

1. Added spaces after the property and method references
2. Changed "8" to "10"
3. Added a sentence to explain where these colors are used in the UI
4. Added the Clear method

This object is represented in the PowerPoint colour picker UI as "Recent Colors" and support 10 (not 8) swatches. I also added the clear method as this is a common request from template designers to be able to tidy up a template by removing the MRU colours prior to distribution, something many currently do by manually hacking the XML of the PowerPoint file.